// lib/modules/segmentor/enums/SegmentOperatorValueEnum.ts
export enum SegmentOperatorValueEnum {
  AND = 'and',
  NOT = 'not',
  OR = 'or',
  CUSTOM_VARIABLE = 'custom_variable',
  USER = 'user',
  COUNTRY = 'country',
  REGION = 'region',
  CITY = 'city',
  OPERATING_SYSTEM = 'os',
  DEVICE_TYPE = 'device_type',
  BROWSER_AGENT = 'browser_string',
  UA = 'ua',
  FEATURE_ID = 'featureId'
}